Add results view to admin polls, extract shared PollResultsDisplay component
- Extract PollResultsDisplay into shared component at src/components/PollResultsDisplay.tsx
- Add 'Results' toggle button per poll in admin list, expands to show live results
- Add PollRow component to Admin.tsx with edit/delete/results actions
- Add bg-background to small inputs in AdminPollForm for consistent appearance
